backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
requests>=2.31.0
pydantic>=2.8.0
python-dotenv>=1.0.1
langchain>=0.3.0
langchain-google-genai>=2.0.0